rf(lAnalytics): 1 file per Metric
