+++ libs/feature/learning-analytics/src/index.ts
@@ -1,1 +1,3 @@
+// export * from "./lib/learning-analytics";
+export * from "./lib/metrics/metrics";
 export * from "./lib/learning-analytics";
